type session manager
